For #5: ARC suggestions after review.
